+++ core/src/test/java/org/elasticsearch/index/mapper/DocumentMapperMergeTests.java
@@ -24,17 +24,8 @@ import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.compress.CompressedXContent;
 import org.elasticsearch.common.xcontent.XContentFactory;
 import org.elasticsearch.common.xcontent.XContentType;
-import org.elasticsearch.index.IndexService;
 import org.elasticsearch.index.analysis.NamedAnalyzer;
-import org.elasticsearch.index.mapper.DocumentFieldMappers;
-import org.elasticsearch.index.mapper.DocumentMapper;
-import org.elasticsearch.index.mapper.DocumentMapperParser;
-import org.elasticsearch.index.mapper.MapperService;
-import org.elasticsearch.index.mapper.Mapping;
-import org.elasticsearch.index.mapper.ObjectMapper;
-import org.elasticsearch.index.mapper.ParsedDocument;
 import org.elasticsearch.test.ESSingleNodeTestCase;
-import org.junit.Before;
 
 import java.io.IOException;
 import java.util.concurrent.CyclicBarrier;
@@ -48,22 +39,12 @@ import static org.hamcrest.Matchers.nullValue;
 
 public class DocumentMapperMergeTests extends ESSingleNodeTestCase {
 
-    IndexService indexService;
-    MapperService mapperService;
-    DocumentMapperParser parser;
-
-    @Before
-    public void setup() {
-        indexService = createIndex("test");
-        mapperService = indexService.mapperService();
-        parser = mapperService.documentMapperParser();
-    }
-
     public void test1Merge() throws Exception {
 
         String stage1Mapping = XContentFactory.jsonBuilder().startObject().startObject("person").startObject("properties")
                 .startObject("name").field("type", "text").endObject()
                 .endObject().endObject().endObject().string();
+        DocumentMapperParser parser = createIndex("test").mapperService().documentMapperParser();
         DocumentMapper stage1 = parser.parse("person", new CompressedXContent(stage1Mapping));
         String stage2Mapping = XContentFactory.jsonBuilder().startObject().startObject("person").startObject("properties")
                 .startObject("name").field("type", "text").endObject()
@@ -82,6 +63,7 @@ public class DocumentMapperMergeTests extends ESSingleNodeTestCase {
     }
 
     public void testMergeObjectDynamic() throws Exception {
+        DocumentMapperParser parser = createIndex("test").mapperService().documentMapperParser();
         String objectMapping = XContentFactory.jsonBuilder().startObject().startObject("type1").endObject().endObject().string();
         DocumentMapper mapper = parser.parse("type1", new CompressedXContent(objectMapping));
         assertNull(mapper.root().dynamic());
@@ -95,6 +77,7 @@ public class DocumentMapperMergeTests extends ESSingleNodeTestCase {
     }
 
     public void testMergeObjectAndNested() throws Exception {
+        DocumentMapperParser parser = createIndex("test").mapperService().documentMapperParser();
         String objectMapping = XContentFactory.jsonBuilder().startObject().startObject("type1").startObject("properties")
                 .startObject("obj").field("type", "object").endObject()
                 .endObject().endObject().endObject().string();
@@ -120,6 +103,7 @@ public class DocumentMapperMergeTests extends ESSingleNodeTestCase {
     }
 
     public void testMergeSearchAnalyzer() throws Exception {
+        DocumentMapperParser parser = createIndex("test").mapperService().documentMapperParser();
         String mapping1 = XContentFactory.jsonBuilder().startObject().startObject("type")
                 .startObject("properties").startObject("field").field("type", "text").field("analyzer", "standard").field("search_analyzer", "whitespace").endObject().endObject()
                 .endObject().endObject().string();
@@ -137,6 +121,7 @@ public class DocumentMapperMergeTests extends ESSingleNodeTestCase {
     }
 
     public void testChangeSearchAnalyzerToDefault() throws Exception {
+        MapperService mapperService = createIndex("test").mapperService();
         String mapping1 = XContentFactory.jsonBuilder().startObject().startObject("type")
                 .startObject("properties").startObject("field").field("type", "text").field("analyzer", "standard").field("search_analyzer", "whitespace").endObject().endObject()
                 .endObject().endObject().string();
@@ -153,6 +138,7 @@ public class DocumentMapperMergeTests extends ESSingleNodeTestCase {
     }
 
     public void testConcurrentMergeTest() throws Throwable {
+        final MapperService mapperService = createIndex("test").mapperService();
         mapperService.merge("test", new CompressedXContent("{\"test\":{}}"), MapperService.MergeReason.MAPPING_UPDATE, false);
         final DocumentMapper documentMapper = mapperService.documentMapper("test");
 
@@ -226,6 +212,7 @@ public class DocumentMapperMergeTests extends ESSingleNodeTestCase {
                         .field("enabled", false)
                     .endObject()
                 .endObject().endObject().bytes());
+        MapperService mapperService = createIndex("test").mapperService();
         mapperService.merge("type", mapping, MapperService.MergeReason.MAPPING_UPDATE, false);
 
         CompressedXContent update = new CompressedXContent(XContentFactory.jsonBuilder().startObject()
@@ -242,7 +229,9 @@ public class DocumentMapperMergeTests extends ESSingleNodeTestCase {
         assertFalse(mapper.sourceMapper().enabled());
     }
 
-    public void testMergeParent() throws IOException {
+    public void testMergeChildType() throws IOException {
+        DocumentMapperParser parser = createIndex("test").mapperService().documentMapperParser();
+
         String initMapping = XContentFactory.jsonBuilder().startObject().startObject("child")
             .startObject("_parent").field("type", "parent").endObject()
             .endObject().endObject().string();
@@ -281,6 +270,8 @@ public class DocumentMapperMergeTests extends ESSingleNodeTestCase {
     }
 
     public void testMergeAddingParent() throws IOException {
+        DocumentMapperParser parser = createIndex("test").mapperService().documentMapperParser();
+
         String initMapping = XContentFactory.jsonBuilder().startObject().startObject("cowboy")
             .startObject("properties")
             .startObject("name").field("type", "text").endObject()
@@ -297,25 +288,5 @@ public class DocumentMapperMergeTests extends ESSingleNodeTestCase {
         DocumentMapper updatedMapper = parser.parse("cowboy", new CompressedXContent(updatedMapping));
         Exception e = expectThrows(IllegalArgumentException.class, () -> initMapper.merge(updatedMapper.mapping(), false));
         assertThat(e.getMessage(), containsString("The _parent field's type option can't be changed: [null]->[parent]"));
-    }
-
-    public void testMergeDate() throws IOException {
-        String initMapping = XContentFactory.jsonBuilder().startObject().startObject("movie")
-            .startObject("properties")
-            .startObject("release_date").field("type", "date").field("format", "yyyy/MM/dd").endObject()
-            .endObject().endObject().endObject().string();
-        DocumentMapper initMapper = parser.parse("movie", new CompressedXContent(initMapping));
-
-        assertThat(initMapper.mappers().getMapper("release_date"), notNullValue());
-        assertFalse(initMapper.mappers().getMapper("release_date").fieldType().stored());
-
-        String updateFormatMapping = XContentFactory.jsonBuilder().startObject().startObject("movie")
-            .startObject("properties")
-            .startObject("release_date").field("type", "date").field("format", "epoch_millis").endObject()
-            .endObject().endObject().endObject().string();
-        DocumentMapper updateFormatMapper = parser.parse("movie", new CompressedXContent(updateFormatMapping));
-
-        Exception e = expectThrows(IllegalArgumentException.class, () -> initMapper.merge(updateFormatMapper.mapping(), false));
-        assertThat(e.getMessage(), containsString("date field's format cannot be updated"));
     }
 }
